En cours - Validation sur l'email (regex) terminée. Il faut maintenant passer à la validation sur les autres champs.
