--- src/main/java/com/projet/validator/InscriptionValidator.java
@@ -1,28 +1,0 @@
-package com.projet.validator;
-
-import javax.faces.bean.ManagedBean;
-import javax.faces.bean.RequestScoped;
-import javax.faces.component.UIComponent;
-import javax.faces.context.FacesContext;
-import javax.faces.validator.FacesValidator;
-import javax.faces.validator.Validator;
-import javax.faces.validator.ValidatorException;
-
-@ManagedBean
-@RequestScoped
-@FacesValidator
-public class InscriptionValidator implements Validator {
-
-	private String email;
-
-	public InscriptionValidator() {
-		super();
-	}
-
-	@Override
-	public void validate(FacesContext context, UIComponent component, Object value) throws ValidatorException {
-				
-	}
-	
-	
-}
